number int -> double

+++ OOP_Exercise/Number.h
@@ -1,4 +1,4 @@
 #ifndef NUMBER_H
 #define NUMBER_H
-typedef int number;
+typedef double number;
 #endif // NUMBER_H
